Refs #90980, fix the namespace of entities and validators for HbxEnrollment
