+++ spec/domain/entities/hbx_enrollments/hbx_enrollment_member_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe Entities::HbxEnrollmentMember, dbclean: :after_each do
+RSpec.describe Entities::HbxEnrollments::HbxEnrollmentMember, dbclean: :after_each do
 
   let(:valid_params) do
     { applicant_id: BSON::ObjectId.new,
@@ -11,7 +11,7 @@ RSpec.describe Entities::HbxEnrollmentMember, dbclean: :after_each do
       coverage_start_on: TimeKeeper.date_of_record}
   end
 
-  let(:contract) { Validators::HbxEnrollmentMemberContract.new }
+  let(:contract) { Validators::HbxEnrollments::HbxEnrollmentMemberContract.new }
 
   it 'contract validation should pass' do
     expect(contract.call(valid_params).to_h).to eq valid_params
@@ -23,7 +23,7 @@ RSpec.describe Entities::HbxEnrollmentMember, dbclean: :after_each do
     end
 
     it 'should return an object' do
-      expect(@result).to be_a(Entities::HbxEnrollmentMember)
+      expect(@result).to be_a(Entities::HbxEnrollments::HbxEnrollmentMember)
     end
   end
